refactor: 완료버튼 로직 분리
