+++ src/components/writePage/index.ts
@@ -1,4 +1,4 @@
 export * from './FileSlider/FileSlider';
 export * from './ReviewForm/ReviewForm';
 export * from './TagContainer/TagContainer';
-export * from './WritePageState/useWritePageState';
+export * from './hooks/useWritePageState';
