+++ src/pages/DetailPage/DetailPage.tsx
@@ -135,6 +135,7 @@ export const DetailPage = () => {
           <StDetailPageContainer>
             {data && (
               <>
+                <button onClick={onEditClickHandler}>임시</button>
                 <StDetailPageHeader>
                   <img
                     src={data.profileImgUrl || noUser}
